restructured code to put security support in its own project
