+++ apiserver/SecuringImplementation.md
@@ -7,11 +7,11 @@ The solution is based on the architecture below:
 
 It basically adds Authorization sidecar to the KubeRay API server pod. This architecture is extremely flexible and allows users to plug sidecar implementations that adhere to their security requirements, that can differ significantly across multiple organizations.
 
-Here we will use a very simple [sidecar implementation](cmd/proxy/proxy.go) a reverse proxy using token based authorization. This is a very simple authorization based on the string token, shared between proxy and client. This implementation is not meant for production, but rather is here as a demonstration for the overall implementation. Additional examples of reverse proxy implementation can be found [here](https://github.com/blublinsky/auth-reverse-proxy). There is also a wealth of open source implementations, for example, [oauth2-proxy](https://github.com/oauth2-proxy/oauth2-proxy) and many commercial offerings.
+Here we will use a very simple [sidecar implementation](../security/cmd/main.go) a reverse proxy using token based authorization. This is a very simple authorization based on the string token, shared between proxy and client. This implementation is not meant for production, but rather is here as a demonstration for the overall implementation. Additional examples of reverse proxy implementation can be found [here](https://github.com/blublinsky/auth-reverse-proxy). There is also a wealth of open source implementations, for example, [oauth2-proxy](https://github.com/oauth2-proxy/oauth2-proxy) and many commercial offerings.
 
 ## Basic token-based authentication reverse proxy
 
-A simple token-based authentication reverse proxy [implementation](cmd/proxy/proxy.go) is provided in the project along with make targets to build a docker image for it and push this image to the image repository. There is also a pre-build image `quay.io/ibmdpdev/api-server-proxy:secure` that you can use for experimenting with security.
+A simple token-based authentication reverse proxy [implementation](../security/cmd/main.go) is provided in the project along with make targets to build a docker image for it and push this image to the image repository. There is also a pre-build image `quay.io/ibmdpdev/api-server-proxy:secure` that you can use for experimenting with security.
 
 ## Installation
