+++ security/go.mod
@@ -1,0 +1,18 @@
+module github.com/ray-project/kuberay/security
+
+go 1.19
+
+require (
+	google.golang.org/grpc v1.59.0
+	google.golang.org/protobuf v1.31.0
+	k8s.io/klog/v2 v2.110.1
+)
+
+require (
+	github.com/go-logr/logr v1.3.0 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	golang.org/x/net v0.14.0 // indirect
+	golang.org/x/sys v0.11.0 // indirect
+	golang.org/x/text v0.12.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20230822172742-b8732ec3820d // indirect
+)
